Not sure of the changes, this commit is prior to changing the index to add the staffing number to the total building capacity.
Added in this commit:
1.) includes/set_term/setTerm.php allows for new terms to be pulled for this report.
2.) includes/set_term/provideTERMAUTO.php allows for the current term to be
determined based on the current date in the calendar. (This was also used in
PETS but tweaked to work with the $myTERM variable inside index.php)

Other changes:
New class for reading an Oracle Query is provided in classes/queryRead.php

read_data_testing.php is stated as a completely new file. This file aggregates
the information provided in the 3 views from John. Building assignments, Students
Staff (RAs) Allotment and Building Capacity.